+++ src/Entries/Collection.php
@@ -2,6 +2,7 @@
 
 namespace Eloquenty\Entries;
 
+use Eloquenty\Facades\EloquentyEntry as EntryFacade;
 use Statamic\Entries\Collection as StatamicCollection;
 
 class Collection extends StatamicCollection
@@ -9,6 +10,6 @@ class Collection extends StatamicCollection
     // Eloquenty: Added for retrieving the Eloquenty QueryBuilder
     public function queryEloquentyEntries()
     {
-        return app(EntryRepository::class)->query()->where('collection', $this->handle());
+        return EntryFacade::query()->where('collection', $this->handle());
     }
 }